Fixed controls

+++ src/Ball.java
@@ -2,7 +2,7 @@ import java.awt.Color;
 import java.awt.geom.Point2D;
 
 public class Ball {
-    public static final int BALL_RADIUS = 12;
+    public static final float BALL_RADIUS = 12;
     private int number;
     private Color color;
     private Point2D location;
